+++ src/api/controllers/deployController/deployInstance.ts
@@ -1,6 +1,6 @@
 import createError from "http-errors";
 
-import createDeploymentInstance from "../../../deploy/build-utils/createDeploymentInstance";
+import createDeploymentInstance from "../../../services/deploy/build-utils/createDeploymentInstance";
 import { createRepoWebhook, getCommits } from "../../github/client";
 
 import catchAsync from "../../../utils/asyncHandler";
