+++ src/api/controllers/updateController/deleteDeployment.ts
@@ -3,10 +3,10 @@ import { startSession } from "mongoose";
 import Config from "../../../config";
 import catchAsync from "../../../utils/asyncHandler";
 
-import terminateInstance from "../../../deploy/aws/ec2_terminateinstances";
-import changeDNSRecord from "../../../deploy/aws/route53_changerecord";
-import describeInstanceIp from "../../../deploy/aws/ec2_describeinstances";
-import deleteLogStream from "../../../deploy/aws/cwl_deletelogstream";
+import terminateInstance from "../../../services/deploy/aws/ec2_terminateinstances";
+import changeDNSRecord from "../../../services/deploy/aws/route53_changerecord";
+import describeInstanceIp from "../../../services/deploy/aws/ec2_describeinstances";
+import deleteLogStream from "../../../services/deploy/aws/cwl_deletelogstream";
 import { deleteRepoWebhook } from "../../github/client";
 
 import { User } from "../../../models/User";
