+++ src/services/deploy/aws/config/setNginxScript.ts
