+++ src/services/deploy/aws/ec2_terminateinstances.ts
@@ -1,10 +1,10 @@
 import { TerminateInstancesCommand } from "@aws-sdk/client-ec2";
 
-import Config from "../../config";
+import Config from "../../../config";
 import ec2Client from "./libs/ec2Client";
 
-import { CustomError } from "../../utils/errors";
-import { createGeneralLogDebug } from "../../utils/createDebug";
+import { CustomError } from "../../../utils/errors";
+import { createGeneralLogDebug } from "../../../utils/createDebug";
 
 const terminateInstance = async (instanceId: string) => {
   const debug = createGeneralLogDebug(Config.CLIENT_OPTIONS.debug);
