+++ src/services/deploy/aws/libs/cloudWatchLogsClient.ts
@@ -1,6 +1,6 @@
 import { CloudWatchLogsClient } from "@aws-sdk/client-cloudwatch-logs";
 
-import Config from "../../../config";
+import Config from "../../../../config";
 
 const REGION = Config.INSTANCE_REGION;
