+++ src/services/deploy/aws/libs/route53Client.ts
@@ -1,6 +1,6 @@
 import { Route53Client } from "@aws-sdk/client-route-53";
 
-import Config from "../../../config";
+import Config from "../../../../config";
 
 const REGION = Config.INSTANCE_REGION;
