+++ src/services/deploy/build-utils/runCertbot.ts
@@ -1,17 +1,17 @@
 import { InstanceStateName } from "@aws-sdk/client-ec2";
 import { RRType } from "@aws-sdk/client-route-53";
 
-import Config from "../../config";
+import Config from "../../../config";
 import runCertbotCommands from "../cli/runCertbotCommands";
 import getRecordInstanceStatus from "./getRecordInstanceStatus";
 import terminateInstance from "../aws/ec2_terminateinstances";
 import changeDNSRecord from "../aws/route53_changerecord";
-import { deleteRepoWebhook } from "../../api/github/client";
+import { deleteRepoWebhook } from "../../../api/github/client";
 
-import { createCertbotDebug } from "../../utils/createDebug";
-import { DeploymentError } from "../../utils/errors";
+import { createCertbotDebug } from "../../../utils/createDebug";
+import { DeploymentError } from "../../../utils/errors";
 
-import { DeploymentData, RecordInstaceStatus } from "../../types/custom";
+import { DeploymentData, RecordInstaceStatus } from "../../../types/custom";
 
 export default async function runCertbot(
   instanceId: string,
